backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
asyncpg==0.23.0
tortoise-orm==0.17.7
aerich==0.5.8
pytest==6.2.5
requests==2.26.0
aiohttp==3.7.4
python-dotenv==0.19.0